Refactor logger access in SqsHostedService
The private `logger` field in `SqsHostedService` was replaced with a protected `Logger` property. This change allows derived classes to access the logger. All previous usages of the `logger` field have been updated to use the new `Logger` property. The active profile line in the `aws.xml` file was also removed, as it is no longer necessary.