Add GetQueueLengthAsync method to ISqsQueueHelper
A new method, GetQueueLengthAsync, was added to the interface ISqsQueueHelper and implemented in the class DefaultSqsQueueHelper. The function retrieves the approximate number of messages in the specified queue, which can be useful for monitoring the queue status. Also, a change was made to the aws.xml to include the activeProfile tag, and a variable visibility changed in SqsMessagePump class from public to private.