Update package versions and project version
Incremented the version of the project from 2.0.1 to 2.0.2. In addition, versions of the Roslynator.Analyzers, Roslynator.CodeAnalysis.Analyzers, Roslynator.Formatting.Analyzers, and AWSSDK.SQS packages were also updated. This update ensures that the latest features, bug fixes, and performance improvements offered by these updated packages are incorporated into the project. The NetEscapades.EnumGenerators package was removed as it was no longer required in the project.