Fix serialization bug and add batch size option in SqsDispatcher
Several changes have been done to enhance SQS message dispatching.
First, a bug was fixed in object serialization when queuing. Previously, the entire object array was being serialized for each index. Now, only the respective object at an index is serialized.
Second, a parameter was added to allow config of the maximum number of messages sent in a batch. This replaces the hard-coded value set to 10 which is the AWS default. The param allows flexibility if AWS changes its default value in future.