Refactor expiry handling in SqsMessagePump
Refactored the way expired messages are handled in SqsMessagePump. Now these messages properly trigger a 'TotalExpired' counter, rather than being added to 'TotalFailures'. Also, the 'IsMessageExpired' method now accepts a 'tags' parameter to add failure tags whenever a message is expired. This was done for better tracking and categorization of expired and deleted messages in the system.