Update and refactor library for greater extensibility and telemetry support
This commit encompasses several changes to the SQS.Extensions library. First, the build version is updated to 2.4.0 and an AWSSDK.SQS dependency is updated for better compatibility and security. Second, the library is refactored for greater extensibility with the DefaultSqsQueueHelper now made public and virtual, allowing developers to override and extend it easily.

Additionally, we have included changes to support telemetry. A detailed section covering telemetry and monitoring with Prometheus and Grafana has been added to the README. A sample Grafana dashboard configuration is provided in grafana/dashboard.json to aid developers in setting up monitoring for their applications using SQS.Extensions.

This upgrade is intended to provide developers more control over the library's behavior and to make monitoring of applications easier and more straightforward.