Update var reference, queue tag name, and versions
Updated the variables reference to follow a consistent pattern in SqsHostedService.cs. Renamed MeterTag constant "QueueUrl" to "QueueName" to better reflect its nature. Upgraded the version of the package from 2.4.1 to 2.4.2 and the AWSSDK.SQS package version from 3.7.200.22 to 3.7.200.26 in Directory.Build.props and SQS.Extensions.csproj, respectively. Optimized the number of times TagList was created in SqsMessagePump.cs. This was done to improve readability, consistency and to upgrade to the new version which may have performance improvements and bug fixes.