Add optional custom serialization to Queue methods
This commit adds support for optional custom serialization in both QueueAsync() and QueueBatchAsync() methods of the SqsDispatcher class, allowing more flexibility in how objects are serialized before being sent to SQS. Also, updates the method comments in 'ISqsDispatcher' to reflect changes. Updated the version in 'Directory.Build.props' to denote the introduction of a new feature.