Refactor MessageContext class and adjust SQS dispatcher and message pump
A constructor was introduced to MessageContext, taking in message ID and attributes. The "required" keyword was removed from MessageId and MessageAttributes properties ensuring the correct setting of these properties only via the added constructor. This ensures all instances of MessageContext are initialized properly.

In SqsDispatcher, MessageAttributes are now set. A minor adjustment was made in adding attributes to the request.

SqsMessagePump was updated to use the new constructor of MessageContext. The logic in setting TargetFramework in the project file was reverted.