Initialize test suit in SqsMessagePumpTests
Added initialization of the test suit, `sut.InitAsync()`, in SqsMessagePumpTests. This step is necessary to ensure that the test suit is properly set up before the pumping of messages is tested. It ensures that the tests are not affected by the state of previous tests and remain isolated from each other.