Update SQS message pump tests and refactor fetch logic
The update includes changes to the SQS message pump tests to use mock functions. This was necessary to avoid direct function calls during testing which can interfere with the unit testing process.
Additionally, the fetching logic in the SqsMessagePump.cs is refactored - the number of messages to fetch is now defined as a constant. The fetching logic now scales with the pump, leading to cleaner more efficient code.
The AWSSDK.SQS version has been bumped up to 3.7.200.34 and the assembly version updated to 4.0.3.