Remove SqsMessagePump and introduce BulkMessagePump and SingleMessagePump
Removed the `SqsMessagePump`, implementing a `SingleMessagePump` and `BulkMessagePump` instead. This was done to better separate functionality, and make the code easier to maintain and extend. `SingleMessagePump` is used for processing individual messages, while `BulkMessagePump` is for processing messages in bulk. It also involves renaming and moving some files and classes for better organization and clarity. The `IMessagePump` interface was split into `ISqsMessagePump` for individual message processing and `ISqsBulkMessagePump` for bulk message processing. This change aims to improve the project structure and enhance the separation of concerns.