Optimize bulk message pump and update version
In the BulkMessagePump.cs file, an early return has been added to the BulkMessagePump to avoid unnecessary processing when no messages are received. This optimizes the handling of message pumping when there are no messages present in the queue.

Additionally, the version number in the Directory.Build.props has been incremented from 5.0.1 to 5.0.2.