Update MessageContext attribute type and version bump
The attribute type in MessageContext class within SQS Extensions has been updated from 'MessageAttributeValue' to 'string'. This change simplifies the retrieval and parsing of message attributes by removing the need to access a nested 'StringValue' property. As a consequence, methods 'GetAttributeAsString' and 'GetAttributeAsInt' were updated to directly return the value. This revision also includes a version update in Directory.Build.props from 5.0.2 to 5.0.3.