Refactor MessageContext to take original Message
Replaced multiple parameters in the MessageContext constructor with a single parameter for the original Message. This optimization simplifies the retrieval of message attributes, reducing the need for manually handling each attribute. Additionally, fetched all attributes of the message in BulkMessagePump by adding "All" to MessageAttributeNames. The version has also been updated from 5.0.3 to 5.0.4. Changes were applied wherever MessageContext is instantiated.