Update AWS SDK, xUnit versions and enhance message fetching process
The AWS SDK version for SQS services is updated to 3.7.200.42 to ensure we benefit from the latest stability improvements and features. The xUnit testing framework version has been bumped to 2.5.1 to keep the unit tests updated.

The constant for the number of messages to fetch has been replaced by a variable which is dynamically set based on the configuration. This allows for more efficient handling of the fetch process as it can now easily be set based on the maximum concurrent operations supported by the configuration.

The solution's version has been incremented to 5.3.0 following these changes due to the major improvement in the message fetching process.